# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Autenticação, instituições e usuários com múltiplos perfis.",
        "title": "basis-avalia API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/relatorios/desempenho": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "relatorios"
                ],
                "summary": "Relatório de desempenho por item de plano",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
        "/api/v1/relatorios/desempenho/exportacao": {
            "get": {
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "relatorios"
                ],
                "summary": "Exportar o relatório de desempenho em CSV",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
        "/api/v1/relatorios/desempenho/por-curso": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "relatorios"
                ],
                "summary": "Cumprimento de metas agregado por curso (gráfico)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.erroResposta": {
            "type": "object",
            "properties": {
                "campo": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}